refactor: createMailTerms and createMailContracts now in proper files
