+++ libs/contract/src/lib/contract/+state/contract.model.ts
@@ -1,11 +1,10 @@
 
 import { createDocumentMeta } from "@blockframes/utils/models-meta";
-import { createMailTerm, Duration } from '../../term/+state/term.model';
+import {   Duration } from '../../term/+state/term.firestore';
 import { Timestamp } from "@blockframes/utils/common-interfaces/timestamp";
 import { toDate } from "@blockframes/utils/helpers";
 import { Contract, Holdback, Mandate, Sale } from "./contract.firestore";
 export { contractStatus, ContractStatus, Holdback, Contract, Mandate, Sale, ContractDocument } from './contract.firestore';
-import { BucketContract } from "@blockframes/contract/bucket/+state";
 
 export function createHoldback(params: Partial<Holdback<Date>> = {}): Holdback {
   return {
@@ -68,12 +67,3 @@ export function convertDuration(duration: Duration<Date | Timestamp>): Duration<
     to: toDate(duration.to),
   }
 }
-
-export function createMailContract(contracts:BucketContract<Timestamp>[]){
-  return contracts.map(contract => ({
-    ...contract,
-    terms: createMailTerm(contract.terms)
-  }));
-}
-
-export type MailContract = ReturnType<typeof createMailContract>[number]
